Cuckoo Filter Criterion benchmarks
My naïve cuckoo implementation consistently outperforms bloom filters
for d>=8. This is great, because with this thing we're aiming for very
low false positive rates, since we are planning to store the filters on
disk.

Of course, this measurement can't be easily translated into the 100k
bloom filters scenario, because we'd still be computing only a constant
number of hash functions, and the larger amount of data processed might
dominate the outcome.

On the other hand, in such a case the number of I/O operations will
dominate everything from hash computations to actual data processing, so
the battle will be between 16 I/O ops for 100k bits (bloom) vs 2 I/O
operations for 16 * 100k bits (cuckoo). The latter reads more data, the
former, reads more chunks.