Add Growable Cuckoo filter
When the cuckoo filter becomes too full, measured by a long chain of
unsuccessful evictions, instead of failing the insert we just add one
more logical entry per bucket, and proceed with a larger filter.

To make this possible we don't maintain a single array of entries, but
each bucket is a separate `Vec<Fingerprint>`, with the associated
overhead.

The main reason is that we don't intend to use this as a stand-alone
filter, but as a COBS-style group of filters. Alternatively, we can
consider this a builder, and `finalize()` in the end into another filter.