Make it impossible to add duplicate keys to Cuckoo Filters
Our implementation did not correctly consider the case that its key is
currently accessible from the alternate bucket on insert, and
unconditionally attempted to insert into the first bucket. That leads to
having the same fingerprint in both of its associated buckets, which in
turn gives a wrong item count.

In such a case, eviction would have "resolved" the deduplication, but
nonetheless this is undesired behavior in a non-counting cuckoo filter.

We change the `Filter::insert(key)` signature to

```rust
Filter::insert(key) -> InsertResult
```

to express things that can potentially happen on insert.