Fix hash function correlation leading to increased fp rates
The `fingerprint` function now uses a different has initialization key,
leading to vastly improved fp rates in certain edge cases.

I'm not sure I fully understand the actual root cause, but some
measurements indicated that e.g. for 4000 buckets, the fp rate was 15x
what it should be, whereas for 3999 it was within the expectations.