Use rayon to run queries concurrently
This gives really interesting performance where on a "small" data set,
we scale by a factor of 6x (linear scaling with 8 cores + reduced clock speed
per core due to CPU power envelope), whereas for a "large" data set, the
actual speedup from concurrently running the queries is 2x.

It seems that we spend way more time in "system" (50% instead of 25%),
so it's probably the I/O that is producing most of the extra work.