Use sync I/O for `load_bucket()`
This increases performance of the query benchmark considerably, almost
to parity with sync code all the way through.

I'm leaving this topic for now - it doesn't seem that I can make the
system any faster using async I/O, which may be related to the nature of
a very fast SSD and page caches. My hunch is that with a blob storage
layer with 100ms of latency, this should really help, a lot.