Run query benchmark for a given duration, not a fixed number of times
Query performance varies a lot depending on the size of each bucket,
which in turn depends on the number of partitions, the number of
elements per partition and the number of buckets to distribute the data
over, so our benchmark can go anywhere from 100k qps to 150qps. To avoid
extremely long benchmark running times and too-short execution for other
configs, run each query for a given amount of time instead.